refactor: move builder to validators
