+++ src/main/factories/http/axios-http-client-factory.ts
@@ -1,4 +1,5 @@
-import { AxiosHttpClient } from '@/infra/http/axios-http-client/axios-http-client'
+import { AxiosHttpClient } from '@/infra/http/axios-htttp-client/axios.http.client'
+
 
 export const makeAxiosHttpClient = (): AxiosHttpClient => {
   return new AxiosHttpClient()
