+++ src/main/factories/usecases/authentication/remote-authentication-factory.ts
@@ -1,6 +1,6 @@
 import { makeApiUrl } from '@/main/factories/http/api-url-factory'
 import { makeAxiosHttpClient } from '@/main/factories/http/axios-http-client-factory'
-import { RemoteAuthentication } from '@/data/usecases/authentication/remote-authentication'
+import { RemoteAuthentication } from '@/data/usecase/authentication/remote-authentication'
 import { Authentication } from '@/domain/usecases'
 
 export const makeRemoteAuthentication = (): Authentication => {
