+++ src/validation/validators/validation-composite/validation-composite.spec.ts
@@ -1,4 +1,4 @@
-import { FieldValidationSpy } from '../test/'
+import { FieldValidationSpy } from '../../test'
 import { ValidationComposite } from './validation-composite'
 import faker from 'faker'
